Clarify hit reaction table indexing

+++ include/main/objHitReact.h
@@ -4,7 +4,7 @@
 #include "game/objects/object_fwd.h"
 #include "main/objHitReact_types.h"
 
-u8 ObjHitReact_Update(GameObject* obj,ObjHitReactEntry *reactionEntryTable,u32 reactionEntryCount,
-                      u32 reactionState,float *reactionStepScale);
+u32 ObjHitReact_Update(GameObject* obj, ObjHitReactEntry* reactionEntryTable, u32 reactionEntryCount, u32 reactionState,
+                       float* reactionStepScale);
 
 #endif /* MAIN_OBJHITREACT_H_ */
